+++ packages/api/src/objectSet/ObjectSetListener.ts
@@ -18,18 +18,21 @@ import type {
   ObjectOrInterfaceDefinition,
   PropertyKeys,
 } from "../ontology/ObjectOrInterface.js";
+import type { CompileTimeMetadata } from "../ontology/ObjectTypeDefinition.js";
+import type { WirePropertyTypes } from "../ontology/WirePropertyTypes.js";
 import type { Osdk } from "../OsdkObjectFrom.js";
 
 export namespace ObjectSetSubscription {
   export interface Listener<
     O extends ObjectOrInterfaceDefinition,
     P extends PropertyKeys<O> = PropertyKeys<O>,
+    R extends boolean = false,
   > {
     /**
      * Specific objects have changed and can be immediately updated
      */
     onChange?: (
-      objectUpdate: ObjectUpdate<O, P>,
+      objectUpdate: ObjectUpdate<O, P, R>,
     ) => void;
 
     /**
@@ -52,21 +55,55 @@ export namespace ObjectSetSubscription {
   /**
    * Options for subscribing to an ObjectSet.
    *
-   * properties - The properties to request a subscription for. Requesting specific properties limits the possible properties
+   * @param properties The properties to request a subscription for. Requesting specific properties limits the possible properties
    * that can be returned from the subscription. If not provided, all properties will be requested and potentially be returned on updates.
+   *
+   * @param includeRid Whether to include the $rid property in the subscription. Defaults to false. RIDs will be returned on all updates unless the update
+   * contains a new value for a geotime series reference property, in which case the RID will be undefined.
    */
   export interface Options<
     O extends ObjectOrInterfaceDefinition,
     P extends PropertyKeys<O> = PropertyKeys<O>,
+    R extends boolean = false,
   > {
+    /**
+     * The properties to request a subscription for. Requesting specific properties limits the possible properties
+     * that can be returned from the subscription. If not provided, all properties will be requested and potentially be returned on updates.
+     */
     properties?: Array<P>;
+
+    /**
+     * Whether to include the $rid property in the subscription. Defaults to false. RIDs will be returned on all updates unless the update
+     * contains a new value for a geotime series reference property, in which case the RID will be undefined. RIDs will not be included
+     * on the objects themselves.
+     */
+    includeRid?: AllFalse<
+      PropertyTypesOnDefMatchesType<O, P, "geotimeSeriesReference">
+    > extends true ? R
+      : false;
   }
 }
 
 type ObjectUpdate<
   O extends ObjectOrInterfaceDefinition,
   P extends PropertyKeys<O>,
+  R extends boolean = false,
 > = {
-  object: Osdk.Instance<O, never, P>;
+  object: R extends false ? Osdk.Instance<O, never, P>
+    : Osdk.Instance<O, "$rid", P>;
   state: "ADDED_OR_UPDATED" | "REMOVED";
 };
+
+type PropertyTypesOnDefMatchesType<
+  Q extends ObjectOrInterfaceDefinition,
+  P extends PropertyKeys<Q>,
+  T extends WirePropertyTypes,
+> = {
+  [K in P]: CompileTimeMetadata<Q>["properties"][K][
+    "type"
+  ] extends T ? true
+    : false;
+};
+
+type AllFalse<T extends Record<string, boolean>> =
+  Exclude<T[keyof T], false> extends never ? true : false;