+++ packages/client/src/object/convertWireToOsdkObjects.test.ts
@@ -79,6 +79,7 @@ describe("convertWireToOsdkObjects", () => {
       "$apiName",
       "$objectType",
       "$primaryKey",
+      "$propertySecurities",
       "$objectSpecifier",
       "employeeLocation",
     ].sort());
@@ -229,6 +230,7 @@ describe("convertWireToOsdkObjects", () => {
       undefined,
       {},
       undefined,
+      undefined,
       false,
     );
     const prototypeAfter = Object.getPrototypeOf(object2);
@@ -260,6 +262,7 @@ describe("convertWireToOsdkObjects", () => {
       {},
       undefined,
       undefined,
+      undefined,
       false,
     );
     const prototypeAfter = Object.getPrototypeOf(object2);
@@ -287,6 +290,7 @@ describe("convertWireToOsdkObjects", () => {
       FooInterface.apiName,
       undefined,
       {},
+      undefined,
     )) as unknown as Osdk<FooInterface>[];
 
     expect(objAsFoo).toMatchInlineSnapshot(`
@@ -295,6 +299,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$title": "Steve",
         "fooSpt": "Steve",
       }
@@ -309,6 +314,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$title": "Steve",
         "employeeId": 0,
         "fullName": "Steve",
@@ -336,6 +342,7 @@ describe("convertWireToOsdkObjects", () => {
       [objectFromWire],
       FooInterface.apiName,
       {},
+      undefined,
       false,
       undefined,
       false,
@@ -348,6 +355,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$title": "Steve",
         "fooIdp": "SEA",
         "fooSpt": "Steve",
@@ -363,6 +371,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$title": "Steve",
         "fullName": "Steve",
         "office": "SEA",
@@ -390,6 +399,7 @@ describe("convertWireToOsdkObjects", () => {
       [objectFromWire],
       FooInterface.apiName,
       {},
+      undefined,
       false,
       undefined,
       false,
@@ -403,6 +413,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$title": "Steve",
         "fooIdp": "SEA",
         "fooSpt": "Steve",
@@ -418,6 +429,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$title": "Steve",
         "fullName": "Steve",
         "office": "SEA",
@@ -446,6 +458,7 @@ describe("convertWireToOsdkObjects", () => {
       FooInterface.apiName,
       undefined,
       {},
+      undefined,
     )) as unknown as Osdk<FooInterface, "$rid" | "$all">[];
 
     expect(objAsFoo).toMatchInlineSnapshot(`
@@ -454,6 +467,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$rid": "hiMom",
         "$title": "Steve",
         "fooSpt": "Steve",
@@ -470,6 +484,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$rid": "hiMom",
         "$title": "Steve",
         "employeeId": 0,
@@ -501,6 +516,7 @@ describe("convertWireToOsdkObjects", () => {
       [objectFromWire],
       FooInterface.apiName,
       {},
+      undefined,
       false,
       undefined,
       false,
@@ -513,6 +529,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$rid": "hiMom",
         "$title": "Steve",
         "fooIdp": "SEA",
@@ -530,6 +547,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$rid": "hiMom",
         "$title": "Steve",
         "employeeId": 0,
@@ -562,6 +580,7 @@ describe("convertWireToOsdkObjects", () => {
       [objectFromWire],
       FooInterface.apiName,
       {},
+      undefined,
       false,
       undefined,
       false,
@@ -575,6 +594,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$rid": "hiMom",
         "$title": "Steve",
         "fooIdp": "SEA",
@@ -592,6 +612,7 @@ describe("convertWireToOsdkObjects", () => {
         "$objectSpecifier": "Employee:0",
         "$objectType": "Employee",
         "$primaryKey": 0,
+        "$propertySecurities": undefined,
         "$rid": "hiMom",
         "$title": "Steve",
         "employeeId": 0,
@@ -616,6 +637,7 @@ describe("convertWireToOsdkObjects", () => {
           undefined,
           undefined,
           {},
+          undefined,
           ["employeeId"],
           "throw",
         )
@@ -637,6 +659,7 @@ describe("convertWireToOsdkObjects", () => {
           undefined,
           undefined,
           {},
+          undefined,
           ["fullName"],
           "throw",
         ),
@@ -655,6 +678,7 @@ describe("convertWireToOsdkObjects", () => {
         undefined,
         undefined,
         {},
+        undefined,
         ["employeeId"],
         "drop",
       );
@@ -674,6 +698,7 @@ describe("convertWireToOsdkObjects", () => {
         undefined,
         undefined,
         {},
+        undefined,
         ["fullName"],
         "drop",
       );
@@ -696,6 +721,7 @@ describe("convertWireToOsdkObjects", () => {
           undefined,
           {},
           undefined,
+          undefined,
           ["employeeId"],
           "throw",
         )
@@ -717,6 +743,7 @@ describe("convertWireToOsdkObjects", () => {
           undefined,
           {},
           undefined,
+          undefined,
           ["fullName"],
           "throw",
         ),
@@ -734,6 +761,7 @@ describe("convertWireToOsdkObjects", () => {
         [object],
         undefined,
         {},
+        undefined,
         undefined,
         ["employeeId"],
         "drop",
@@ -753,6 +781,7 @@ describe("convertWireToOsdkObjects", () => {
         [object],
         undefined,
         {},
+        undefined,
         undefined,
         ["fullName"],
         "drop",
@@ -777,6 +806,7 @@ describe("convertWireToOsdkObjects", () => {
           {},
           undefined,
           undefined,
+          undefined,
           "throw",
         )
       ).rejects.toThrowErrorMatchingInlineSnapshot(
@@ -799,6 +829,7 @@ describe("convertWireToOsdkObjects", () => {
           {},
           undefined,
           undefined,
+          undefined,
           "throw",
         ),
       ).resolves.to.not.toBeUndefined();
@@ -815,6 +846,7 @@ describe("convertWireToOsdkObjects", () => {
         [object],
         undefined,
         {},
+        undefined,
         undefined,
         undefined,
         "drop",
@@ -835,6 +867,7 @@ describe("convertWireToOsdkObjects", () => {
         [object],
         undefined,
         {},
+        undefined,
         undefined,
         undefined,
         "drop",
@@ -859,6 +892,7 @@ describe("convertWireToOsdkObjects", () => {
           {},
           undefined,
           undefined,
+          undefined,
           "throw",
         )
       ).rejects.toThrowErrorMatchingInlineSnapshot(
@@ -881,6 +915,7 @@ describe("convertWireToOsdkObjects", () => {
           {},
           undefined,
           undefined,
+          undefined,
           "throw",
         ),
       ).resolves.to.not.toBeUndefined();
@@ -897,6 +932,7 @@ describe("convertWireToOsdkObjects", () => {
         [object],
         undefined,
         {},
+        undefined,
         undefined,
         undefined,
         "drop",
@@ -919,6 +955,7 @@ describe("convertWireToOsdkObjects", () => {
         {},
         undefined,
         undefined,
+        undefined,
         "drop",
       );
 
@@ -938,6 +975,7 @@ describe("convertWireToOsdkObjects", () => {
       [object],
       "FooInterface",
       {},
+      undefined,
       undefined,
       ["fooSpt"],
       "drop",
@@ -960,6 +998,7 @@ describe("convertWireToOsdkObjects", () => {
       [object],
       "FooInterface",
       {},
+      undefined,
       undefined,
       ["fooSpt", "fooIdp"],
       "drop",
@@ -1071,6 +1110,15 @@ describe("convertWireToOsdkObjects", () => {
                   "multiplicity": false,
                   "nullable": true,
                   "type": "stringTimeseries",
+                  "valueFormatting": undefined,
+                  "valueTypeApiName": undefined,
+                },
+                "favoriteRestaurants": {
+                  "description": undefined,
+                  "displayName": undefined,
+                  "multiplicity": true,
+                  "nullable": true,
+                  "type": "string",
                   "valueFormatting": undefined,
                   "valueTypeApiName": undefined,
                 },